% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Infer the parameters of a (hierarchical) sequential regression
model using Stan.}
\usage{
fit(formula, data, family = c("cumulative", "sratio", "cratio"), ...)
}
\arguments{
\item{formula}{an \code{lme4}-style formula}

\item{data}{a data frame containing the variables specified in \code{forumla}}

\item{family}{name of the family to be used}

\item{...}{additional parameters passed to \code{\link[rstan]{sampling}}}
}
\value{
returns a \code{\link[rstan]{stan.fit}} object
}
\description{
Infers posterior distributions of the parameters
of a sequential regression model using
Stan's dynamic Hamiltonian Monte Carlo sampler.
}
\examples{
data("tonsil")
fit(tonsil ~ carrier, tonsil, "sratio", iter = 1000, chains = 2)

data("wine")
fit(rating ~ temperature + (1 | judge), wine, iter = 1000, chains = 2)

}
