% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{make_standata}
\alias{make_standata}
\title{Data sets for sequential models}
\usage{
make_standata(formula, data, family = c("cumulative", "sratio", "cratio"))
}
\arguments{
\item{formula}{an \code{lme4}-style formula}

\item{data}{a data frame containing the variables specified in \code{forumla}}

\item{family}{name of the family to be used}
}
\value{
returns a \code{\link[rstan]{stan.fit}} object
}
\description{
Generate the model matrices and response vector
required for \code{link{rstan}{sampling}}
}
\examples{
data("tonsil")
fit(tonsil ~ carrier, tonsil, "sratio", iter = 1000, chains = 2)

data("wine")
fit(rating ~ temperature + (1 | judge), wine, iter = 1000, chains = 2)

}
