% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstansequential-package.R
\docType{package}
\name{rstansequential-package}
\alias{rstansequential-package}
\alias{rstansequential}
\title{Hierarchical sequential regression models using Stan}
\description{
The `rstansequential` package fits hiearchical sequential
 regression models using Stan. The package was created for a case study and
 only offers little functionality other than fitting sequential models for
 regression.
}
